socket: get back to 32bit ports
This is what libd2sock does: instead of shortening writes in asm,
this can be solved on C level by making port 32bit in wrappers.

That approach also fixes the same memory corruption, but w/o careful
asm coding.

+++ src/libc/posix/socket/csock/c_accept.c
@@ -45,7 +45,7 @@ __csock_accept (LSCK_SOCKET *lsd, LSCK_SOCKET *nsd,
   LSCK_SOCKET_CSOCK *csock    = (LSCK_SOCKET_CSOCK *) lsd->idata;
   LSCK_SOCKET_CSOCK *ncsock   = NULL;
   int ret                     = -1;
-  unsigned short dest_port    =  0;
+  unsigned int dest_port      =  0;
   unsigned int dest_addr      =  0;
   struct sockaddr_in *addr_sa = (struct sockaddr_in *) addr;
   int blocking                = lsd->blocking;
